+++ tests/sdk/test_consistent_return_shapes.py
@@ -461,7 +461,10 @@ class TestCLI:
         m = re.match(
             r"Workflow Submitted! Run ID: (?P<run_id>.*)", run_ray.stdout.decode()
         )
-        assert m is not None
+
+        assert (
+            m is not None
+        ), f"STDOUT: {run_ray.stdout.decode()},\n\nSTDERR: {run_ray.stderr.decode()}"
         run_id_ray = m.group("run_id").strip()
         assert "Workflow Submitted!" in run_ce.stdout.decode()
 
@@ -520,7 +523,9 @@ class TestCLI:
         m = re.match(
             r"Workflow Submitted! Run ID: (?P<run_id>.*)", run_ray.stdout.decode()
         )
-        assert m is not None
+        assert (
+            m is not None
+        ), f"STDOUT: {run_ray.stdout.decode()},\n\nSTDERR: {run_ray.stderr.decode()}"
         run_id_ray = m.group("run_id").strip()
         assert "Workflow Submitted!" in run_ce.stdout.decode()
 
@@ -591,7 +596,9 @@ class TestCLIDownloadDir:
         m = re.match(
             r"Workflow Submitted! Run ID: (?P<run_id>.*)", run_ray.stdout.decode()
         )
-        assert m is not None
+        assert (
+            m is not None
+        ), f"STDOUT: {run_ray.stdout.decode()},\n\nSTDERR: {run_ray.stderr.decode()}"
         run_id_ray = m.group("run_id").strip()
         assert mock_ce_run_single in run_ce.stdout.decode()
 
@@ -669,7 +676,9 @@ class TestCLIDownloadDir:
         m = re.match(
             r"Workflow Submitted! Run ID: (?P<run_id>.*)", run_ray.stdout.decode()
         )
-        assert m is not None
+        assert (
+            m is not None
+        ), f"STDOUT: {run_ray.stdout.decode()},\n\nSTDERR: {run_ray.stderr.decode()}"
         run_id_ray = m.group("run_id").strip()
         assert mock_ce_run_multiple in run_ce.stdout.decode()